+++ src/refactoring/4_features/coupon/hooks/use-coupons.ts
@@ -3,6 +3,7 @@ import { useAtom } from "jotai";
 import { couponsAtom } from "../../../5_entities/coupon/model/atoms";
 import { Coupon } from "../../../6_shared/types";
 
+// 관리자 페이지 - 쿠폰 관리
 export const useCoupons = () => {
   const [coupons, setCoupons] = useAtom(couponsAtom);
