+++ src/refactoring/5_entities/cart/model/atoms.ts
@@ -1,6 +1,6 @@
 import { atom } from "jotai";
 
-import { CartItem } from "../../../6_shared/types/domain";
+import { CartItem } from "../../../6_shared/types";
 import { calculateCartTotal } from "../lib";
 
 // 장바구니 아이템 목록
@@ -11,13 +11,4 @@ export const cartTotalAtom = atom((get) => {
   const cart = get(cartAtom);
 
   return calculateCartTotal(cart);
-});
-
-// 장바구니 아이템 조회 함수
-export const findCartItemAtom = atom((get) => {
-  const cart = get(cartAtom);
-
-  return (productId: string) => {
-    return cart.find((item) => item.product.id === productId);
-  };
 });
