+++ src/refactoring/5_entities/coupon/model/atoms.ts
@@ -1,8 +1,10 @@
 import { atom } from "jotai";
 
-import { Coupon } from "../../../6_shared/types/domain";
+import { Coupon } from "../../../6_shared/types";
 import { initialCoupons } from "./constants";
 
+// 관리자 페이지 - 쿠폰 목록
 export const couponsAtom = atom<Coupon[]>(initialCoupons);
 
+// 장바구니 페이지 - 선택된 쿠폰
 export const selectedCouponAtom = atom<Coupon | null>(null);
