+++ src/refactoring/5_entities/product/model/atoms.ts
@@ -1,6 +1,6 @@
 import { atom } from "jotai";
 
-import { Product } from "../../../6_shared/types/domain";
+import { Product } from "../../../6_shared/types";
 import { initialProducts } from "./constants";
 
 export const productsAtom = atom<Product[]>(initialProducts);
