--- src/refactoring/5_shared/store/product.ts
@@ -1,1 +1,0 @@
-
