--- src/refactoring/6_shared/types/domain.ts
@@ -1,32 +1,0 @@
-export interface Product {
-  id: string;
-  name: string;
-  price: number;
-  stock: number;
-  discounts: Discount[];
-}
-
-export interface Discount {
-  quantity: number;
-  rate: number;
-}
-
-export interface CartItem {
-  product: Product;
-  quantity: number;
-}
-
-export interface CartTotal {
-  totalBeforeDiscount: number;
-  totalAfterDiscount: number;
-  totalDiscount: number;
-}
-
-export interface Coupon {
-  name: string;
-  code: string;
-  discountType: "amount" | "percentage";
-  discountValue: number;
-}
-
-export interface ProductFormData extends Omit<Product, "id"> {}
